refactor(tracking): lift row conversion and lifecycle errors out of run_database
Refs #685 — the `tracking/run_database.py` entry in the split wave.

Two helpers come away cleanly, and neither is a line-count cut:

- `run_database_rows.py` — `row_to_run` / `row_to_phase_stats` / `_row_value`
  are pure `sqlite3.Row` -> model functions with no reference to
  `RunDatabase`. `cache/database.py` already has exactly this module beside
  it as `cache/database_rows.py`; this mirrors the established seam.
- `run_lifecycle_errors.py` — the two exception types and the two `raise_*`
  helpers that re-read the run to say which precondition a refused
  transition missed. They are one contract, shared by four lifecycle
  methods, and the package already re-exports the exceptions as its public
  error surface.

The reads/writes seam was considered and rejected: four lifecycle writers
return `self.get_run(run_id)`, and `get_run`/`list_runs` call
`get_phase_stats`, so the halves are genuinely entangled and splitting them
would need delegation boilerplate for twelve methods.

run_database.py 850 -> 731 lines (under the 800 soft limit).

The `raise_*` helpers drop their leading underscore now that they are
imported across a module boundary. No test patches any moved symbol —
every `patch("...run_database...")` target is `RunDatabase`, which stays
put — so no test needed rewiring. `update_run_status` (18) and
`update_operational_phase` (21) are untouched; the complexipy snapshot
diff is pure line renumbering with no new, removed, or changed entry.
