+++ complexipy-snapshot.json
@@ -4740,55 +4740,55 @@
       {
         "name": "RunDatabase::update_run_status",
         "complexity": 18,
-        "line_start": 342,
-        "line_end": 403,
+        "line_start": 223,
+        "line_end": 284,
         "line_complexities": [
           {
-            "line": 357,
+            "line": 238,
             "complexity": 0
           },
           {
-            "line": 358,
+            "line": 239,
             "complexity": 0
           },
           {
-            "line": 359,
+            "line": 240,
             "complexity": 0
           },
           {
-            "line": 361,
+            "line": 242,
             "complexity": 2
           },
           {
-            "line": 365,
+            "line": 246,
             "complexity": 2
           },
           {
-            "line": 369,
+            "line": 250,
             "complexity": 2
           },
           {
-            "line": 373,
+            "line": 254,
             "complexity": 2
           },
           {
-            "line": 377,
+            "line": 258,
             "complexity": 2
           },
           {
-            "line": 381,
+            "line": 262,
             "complexity": 2
           },
           {
-            "line": 385,
+            "line": 266,
             "complexity": 2
           },
           {
-            "line": 389,
+            "line": 270,
             "complexity": 2
           },
           {
-            "line": 393,
+            "line": 274,
             "complexity": 2
           }
         ]
@@ -4796,69 +4796,69 @@
       {
         "name": "RunDatabase::update_operational_phase",
         "complexity": 21,
-        "line_start": 252,
-        "line_end": 285,
+        "line_start": 133,
+        "line_end": 166,
         "line_complexities": [
           {
-            "line": 254,
+            "line": 135,
             "complexity": 0
           },
           {
-            "line": 255,
+            "line": 136,
             "complexity": 0
           },
           {
-            "line": 259,
+            "line": 140,
             "complexity": 2
           },
           {
-            "line": 260,
+            "line": 141,
             "complexity": 0
           },
           {
-            "line": 261,
+            "line": 142,
             "complexity": 2
           },
           {
-            "line": 262,
+            "line": 143,
             "complexity": 3
           },
           {
-            "line": 263,
+            "line": 144,
             "complexity": 0
           },
           {
-            "line": 268,
+            "line": 149,
             "complexity": 0
           },
           {
-            "line": 269,
+            "line": 150,
             "complexity": 2
           },
           {
-            "line": 270,
+            "line": 151,
             "complexity": 0
           },
           {
-            "line": 273,
+            "line": 154,
             "complexity": 3
           },
           {
-            "line": 274,
+            "line": 155,
             "complexity": 4
           },
           {
-            "line": 279,
+            "line": 160,
             "complexity": 5
           },
           {
-            "line": 285,
+            "line": 166,
             "complexity": 0
           }
         ]
       }
     ],
-    "complexity": 106
+    "complexity": 101
   },
   {
     "path": "src/immich_memories/ui/pages/step2_review.py",
